Remove start_time fron UI
It is messing up the klines database, since requesting a start_time will
trigger a db cleanup and retrieve new data from binance. In some cases,
this will return so few data it will not displayable in the graph.

Missing this feature does not impact front-end, because plotly does not
seem to have a breakpoint where we can reload data by scrolling the time
series back