Fix take profit order and safety order triggering together
When safety order triggers, because current price is below take profit,
it will trigger SELL when trailling is set.

Instead, reset the take_profit_price and trailling, so that trailling take_profit
is not triggered. It should only trigger trailling when current price
surpasses take_profit and this take_profit is based on the new reduced
average buy_price