Remove ta-lib
ta-lib is mainly used for candlestick patterns, which so far it hasn't
been useful for detecting trends. Either they never trigger or they
trigger so many times most are false positives.

Besides, it also causes a trouble once in a while in the dockerfile. So
it's not worth sticking to it. The main functions used such as standard
deviation and chaikin oscillators can easily be replicated by another
library.