Short sell price adjustments (#389)
* Short sell price adjustments

Previous short sell price was not the lowest, the condition was wrong. Also qfl hodloo panic signals tend to be downtrend, so to avoid big drops, such as 7% decreases we should short sell earlier.

* Update pip.lock

* Fix bot short strategy form

* Fix available balance and add strategy mode in UI

* Refactor autotrade activation order

It's better to execute test autotrade first, because autotrade
activation would be uncessary if there is no balance available. The
problem of blocking autotrade if there are bugs in the test autotrade is
to wrap it with a catch block, and pass if there is an exception

* Increase timeout for candlestick

* hypercorn implementation

* Update python packages

* Add webapp build stage

* Fix real bots not trading because of no funds

* Fix 401 api issues

* Test nginx unit implementation

Co-authored-by: Carlos Wu Fei <carkod@Carloss-MacBook-Pro.local>