Excute stop loss reversal (#475)
* Excute stop loss reversal
* Trailling take_profit feature
* Market domination improvements
* Handle already bought back margin trades
* Loan repayment fixes
* Trailling profit for margin_short
* Fix margin_short reversal and add it to web UI
* Set trailling option as boolean
* Short trailling profit fixes
* Fix base order for web UI
* Handle more missed margin cases
* Fix trailling toggle with new boolean. Add dynamic trailling switch
* Fix trailling stop_loss streaming updates and add market orders
* Hide autoswitch for not margin bots
* Remove condition to only show autoswitch.
There will be in the future an option to autoswitch for long bots, so removing this to avoid further bugs down the line when autoswitch is used.
There is no logic to manipulate this property, therefore, having it will not cause any issues
* Fixes for long trailling web UI chart
* Add button to hide deal information
* Add stop_loss markup on top of base_order_size, so we can cover for losses
Because margin_short bots run out of balance when repaying back an amount that now is more costly when prices go up, we need to cover the loss.
* Refactor and centralize sell orders
Now that the logic is clearer, refactor all sell order into one single sell order function. Instead of using libraries which don't have continuous support or have missing features, I opted to use the custom signed_request function, which is very simple and straight-forward.
On the other hand, the logic is clear - if there is no matching_engine price, then execute market order. This is something not supported by libraries, unique to this application
* Corrections to the bid/ask price