Add endpoint for isolated account check
Did not implement it to Binquant. If isolated margin bot doesn't open because it doesn't exist, it will show up in the Dashboard, this will allow us to check against telegram signals to know tht that those specific signals were not triggered for this reason. Otherwise, signals would be obsfuscated and we wouldn't for what reason they didn't trigger the bot. That's why isolated check has not been implemented in Binquant.