Global replace of usdt -> usdc (#595)
* Global replace of usdt -> usdc

* Remove the need for bot id in the restart streaming producer

This is because not only bots use it, but also autotrade settings which don't have bot ids

* Fix autotrade settings edit

* Fix balance estimation with USDC and improve performance by replacing endpoints

* Add transition fix for balance series data, while we collect more USDC data

* USDC fixes

* Add endpoint for isolated account check

Did not implement it to Binquant. If isolated margin bot doesn't open because it doesn't exist, it will show up in the Dashboard, this will allow us to check against telegram signals to know tht that those specific signals were not triggered for this reason. Otherwise, signals would be obsfuscated and we wouldn't for what reason they didn't trigger the bot. That's why isolated check has not been implemented in Binquant.