Display order timestamp (#603)
* Display order timestamp

To debug potential gaps between order sell execution and stop loss execution. Sometimes, it seems the order doesn't fill immediately, so this is the first step to solve whatever is blocking the margin short switch from long to short strategy

* Adjust bollinguer band updates to make it tighter

Previously, because bollinguer bands update the tp, sl, td constantly, when position is adverse (prices go down for long bots and prices go up for short bots) positions don't close when returns are higly positive (i.e. 10%) because the volatility is so high, that Bollinguer bands already stretched out the tp and tds. As volatility decreases, price returns to previous value, losing all the returns.

For that reason, add conditions to only update the values with bollinguer bands when required, i.e. when prices are going up for long and prices are going down for short.

* Stop bb_spreads from updating stop losses

---------

Co-authored-by: Carlos Wu <carlos.wufei@houseful.com>