Setup new bot data controller (#636)
* Setup new bot data controller

* Simplify database operations

By creating simpler and atomic db operations as a separate layer, we reduce complexity for testing (duplicated code), decouple dependencies of database and services and also reduce circular imports.

* Refactor and unify simulated orders (paper trading)

* Fix tests for new SQL bots migration

* Refactor all data models to use a single BotModel

The reason behind this is that pydantic table models are SQL models which are different from traditional pydantic models. So to reduce repetition, we have to create a BotBase which is inherited by the "guardian" i.e. the BotModel which is use to guard data everywhere, and the separation of database tables i.e. BotTable

* Fix all SQLModel vs Pydantic model compatibility issues
* Refactor deals into a singleton

The DealFactory is a literal factory, not a factory pattern, in that we do know what kind of object we are creating, but we are unifying into a single point of entry (singleton)