Fix deal id integrity error
Integrity errors were raised by ORM, because deal can be created every time. Therefore, always get deal id from db, as there can only be one deal per bot upon creation. And if there is no deal, then do not create in the save crud function, raise an error to suggest the consumer to use the create function. On the other hand, remove id from DealModel, this was the culprit that was resetting id every time a DealModel is filled with new data, it created a new id.