Increase market_domination data (#697)
* Increase amount of market_domination data so AI can analyze (min 193)

It's not possible to integrate market_domination data with candlesticks, so increase market_domination data by increasing the expire after seconds, this will allow data to stay in the DB longer for AI minimum requirements.

Market domination's main data component is the 24-hour ticker. This ticker apparently uses book order data to build up from. Since we don't store such low level data (it'd be a huge database consuming a lot of space), we can't merge with candlesticks.

Candlesticks are an approximation of the aggregated book orders, they are not exactly the same. That's why even on Binance's data, candlesticks and 24-hour tickers never match.

* Fix update of expire after seconds

* Fix expired orders

* Update kline expirations in order to increase amount of stored data

* Fix paper trading autotrade settings

* Fix active pairs endpoint for paper-trading