Db fields quote asset (#733)
* Add new bot quote assets field

Fix base order quote asset updates and round leading zero numbers

Rename round numbers to more generic maths utility

fix btc correlation data

Fix cronjob

Add fiat order size and move base order size

New fiat order size field is used as drop-in replacement for the base order size. Base order size is moved to Deal table and DealModel.
The reason is that to support multiple base assets (BTC, ETH in addition to USDC) but still keeping USDC as the hedge fiat, we need an "intermediary" currency, and that is the quote asset. So for DASHBTC, the "intermediary" would be BTC, basically because DASH is not available in USDC markets. Eventually closing position and getting back USDC.

So we still want to keep USDC as a fiat, the one we use to allocate our resources. BTC however in this example would be used for the rest of the bot's transactions, that's why the one in deal should be base order size, this is the one that will keep changing and being used and reused. fiat base order size would be unchanged for the rest of the bot's life, because it is the initial one that was set and would not change no matter what transactions happen.

This way we still keep with the initial concept of BotModel being the static, the template on which deals are carried out as devised initially during the design of the Bot data structure.

Add new multicurrency base order functionality

Remove exchange info update on initialization

Fix issue with unavailable quote by rebuying quote asset

Update paper trading table with missing fields from bot table and sell quote asset after trailling, stop loss and take profit

Fix paper trading bot models fields

* Fix streaming error not found symbol