Remove deep book price search (#791)
* Remove deep book price search
This is causing base order price to get to the wick rather than the body of the candle. Which eventually means buy order is way out of the buying price and ends up in a loss.
* Revert back changes in book depth
* Fix bot profit negative numbers
* Remove comissions
They were incorrect, because commissions is appended in the back-end directly from API, which means they are in BNB, not in the USDC fiat